split getGlobalValueManager
this way tile-environment doesn't depend on a specific shared model directly
