+++ v3/src/components/slider/component-handler-slider.test.ts
@@ -3,7 +3,8 @@ import { DIComponentInfo } from "../../data-interactive/data-interactive-types"
 import { diComponentHandler } from "../../data-interactive/handlers/component-handler"
 import { testGetComponent } from "../../data-interactive/handlers/component-handler-test-utils"
 import { appState } from "../../models/app-state"
-import { getGlobalValueManager, getSharedModelManager } from "../../models/tiles/tile-environment"
+import { getSharedModelManager } from "../../models/tiles/tile-environment"
+import { getGlobalValueManager } from "../../models/global/global-value-manager"
 import { toV3Id } from "../../utilities/codap-utils"
 import { ISliderModel, isSliderModel } from "./slider-model"
 import { kSliderIdPrefix } from "./slider-registration"
