+++ v3/src/components/slider/slider-model.ts
@@ -6,7 +6,8 @@ import {
 import { GlobalValue } from "../../models/global/global-value"
 import { applyModelChange } from "../../models/history/apply-model-change"
 import { ISharedModel } from "../../models/shared/shared-model"
-import { getGlobalValueManager, getSharedModelManager } from "../../models/tiles/tile-environment"
+import { getSharedModelManager } from "../../models/tiles/tile-environment"
+import { getGlobalValueManager } from "../../models/global/global-value-manager"
 import { ITileContentModel, TileContentModel } from "../../models/tiles/tile-content"
 import { DateUnit, dateUnits, unitsStringToMilliseconds } from "../../utilities/date-utils"
 import { kSliderTileType } from "./slider-defs"
