+++ v3/src/components/slider/slider-registration.test.ts
@@ -4,7 +4,8 @@ import { FreeTileRow, IFreeTileRow } from "../../models/document/free-tile-row"
 import { GlobalValue, IGlobalValueSnapshot } from "../../models/global/global-value"
 import { getTileComponentInfo } from "../../models/tiles/tile-component-info"
 import { getTileContentInfo } from "../../models/tiles/tile-content-info"
-import { getGlobalValueManager, getSharedModelManager } from "../../models/tiles/tile-environment"
+import { getSharedModelManager } from "../../models/tiles/tile-environment"
+import { getGlobalValueManager } from "../../models/global/global-value-manager"
 import { ITileModelSnapshotIn } from "../../models/tiles/tile-model"
 import { toV2Id } from "../../utilities/codap-utils"
 import { CodapV2Document } from "../../v2/codap-v2-document"
