+++ v3/src/components/slider/slider-registration.ts
@@ -4,10 +4,9 @@ import { V2Slider } from "../../data-interactive/data-interactive-component-type
 import { registerComponentHandler } from "../../data-interactive/handlers/component-handler"
 import { errorResult } from "../../data-interactive/handlers/di-results"
 import { appState } from "../../models/app-state"
-import { GlobalValueManager } from "../../models/global/global-value-manager"
+import { GlobalValueManager, getGlobalValueManager } from "../../models/global/global-value-manager"
 import { registerTileComponentInfo } from "../../models/tiles/tile-component-info"
 import { ITileLikeModel, registerTileContentInfo } from "../../models/tiles/tile-content-info"
-import { getGlobalValueManager } from "../../models/tiles/tile-environment"
 import { ITileModelSnapshotIn } from "../../models/tiles/tile-model"
 import { toV2Id, toV3GlobalId, toV3Id } from "../../utilities/codap-utils"
 import { DateUnit, unitsStringToMilliseconds } from "../../utilities/date-utils"
