+++ v3/src/data-interactive/resource-parser-utils.ts
@@ -7,7 +7,8 @@ import { displayToCanonical } from "../models/formula/utils/canonicalization-uti
 import { getDisplayNameMap } from "../models/formula/utils/name-mapping-utils"
 import { getSharedDataSets } from "../models/shared/shared-data-utils"
 import { getTileContentInfo } from "../models/tiles/tile-content-info"
-import { getGlobalValueManager, getSharedModelManager } from "../models/tiles/tile-environment"
+import { getSharedModelManager } from "../models/tiles/tile-environment"
+import { getGlobalValueManager } from "../models/global/global-value-manager"
 import { toV3Id } from "../utilities/codap-utils"
 import { t } from "../utilities/translation/translate"
 import { DIParsedQuery, DIQueryFunction } from "./data-interactive-types"
