+++ v3/src/models/formula/test-utils/formula-test-utils.ts
@@ -1,7 +1,8 @@
 import { boundaryManager } from "../../boundaries/boundary-manager"
 import { createCodapDocument } from "../../codap/create-codap-document"
 import { getSharedDataSets } from "../../shared/shared-data-utils"
-import { getSharedModelManager, getGlobalValueManager } from "../../tiles/tile-environment"
+import { getSharedModelManager } from "../../tiles/tile-environment"
+import { getGlobalValueManager } from "../../global/global-value-manager"
 import { FormulaMathJsScope } from "../formula-mathjs-scope"
 import { math } from "../functions/math"
 import { displayToCanonical } from "../utils/canonicalization-utils"
