+++ v3/src/v2/export-v2-document.test.ts
@@ -4,7 +4,8 @@ import { DataBroker } from "../models/data/data-broker"
 import { DataSet } from "../models/data/data-set"
 import { GlobalValue } from "../models/global/global-value"
 import { getSharedDataSets } from "../models/shared/shared-data-utils"
-import { getGlobalValueManager, getSharedModelManager } from "../models/tiles/tile-environment"
+import { getSharedModelManager } from "../models/tiles/tile-environment"
+import { getGlobalValueManager } from "../models/global/global-value-manager"
 import { toV2Id } from "../utilities/codap-utils"
 import { ICodapV2DataContext } from "./codap-v2-types"
 import { exportV2Document } from "./export-v2-document"
