+++ v3/src/v2/export-v2-document.ts
@@ -4,7 +4,8 @@ import { convertDataSetToV2 } from "../data-interactive/data-interactive-type-ut
 import { IDocumentModel } from "../models/document/document"
 import { isFreeTileRow } from "../models/document/free-tile-row"
 import { getSharedDataSets } from "../models/shared/shared-data-utils"
-import { getGlobalValueManager, getSharedModelManager } from "../models/tiles/tile-environment"
+import { getSharedModelManager } from "../models/tiles/tile-environment"
+import { getGlobalValueManager } from "../models/global/global-value-manager"
 import { toV2Id } from "../utilities/codap-utils"
 import { exportV2Component } from "./codap-v2-tile-exporters"
 import { CodapV2Component, ICodapV2DataContext, ICodapV2DocumentJson } from "./codap-v2-types"
